// src/lib2/models/Schedule.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Schedule = {
    active?: boolean;
    cron?: string;
    id?: string;
    last_run_at?: string;
    next_run_at?: string;
    pipeline_id?: string;
    timezone?: string;
};